sched: window-stats: Fix overflow bug
Multiplication over-flow possibility exists in update_task_ravg() when
updating task's window_start. That would lead to incorrect accounting
of task load. Fix the issue by using 64-bit arithmetic.

CRs-Fixed: 665706

+++ kernel/sched/core.c
@@ -1551,7 +1551,8 @@ void update_task_ravg(struct task_struct *p, struct rq *rq, int update_sum)
 			if (!update_sum)
 				p->ravg.window_start = wallclock;
 			else
-				p->ravg.window_start += n * window_size;
+				p->ravg.window_start += (u64)n *
+							 (u64)window_size;
 			BUG_ON(p->ravg.window_start > wallclock);
 			if (update_sum)
 				update_history(rq, p, window_size, n);
